Move terminal setup and cleanup to `utils` module
